Added Timers
Timers that are accurate with Single Trip time over the internet between client and server. This makes sure they are in sync with each other with a very low degree of inaccuracy. Also added a GameInstance Subsystem to allow persistance of the server parameters used with GameLift, while we travel from a match back to the lobby.